requirements.txt
numpy==2.3.2
pygame-ce==2.5.5
